v0.3.0
### FEATURES

- Add Region
The "Add Region" functionality allows the user to create a fresh region. Each parameter can be changes and, optionally, templates can be updated to reflect the changes on subsequent actions. It includes server-side namespace and file path validation.

The client code is fully custom and it is mostly contained in the file `uiRegionShared.js` to be re-used for the option: Edit Region.

The server test files are prefixed with the "z" to ensure they get executed as last

v0.3.0

### FEATURES

- Add Region
The "Add Region" functionality allows the user to create a fresh region. Each parameter can be changes and, optionally, templates can be updated to reflect the changes on subsequent actions. It includes server-side namespace and file path validation.

The client code is fully custom and it is mostly contained in the file `uiRegionShared.js` to be re-used for the option: Edit Region.

The server test files are prefixed with the "z" to ensure they get executed as last