Test system for YDBGUI
The testing is performed by the means of standard Java Script testing tools: chai and puppeteer.
While chai manages the tests, puppeteer provides a headless Chrome browser with DOM access, to inspect graphic elements.

We test Client and Server independently, in the following way:

The Client testing is performed by generating different mock data in the `app.system` object, thus emulating multiple server responses scenario's. Each test number generates a different response, stimulating the interface to react according, so that we can inspect the graphic elements and determine the validity of the test.

The Server testing is not using the interface at all and is fully focused on the response returned by the REST call. By modifying the database structure, using OS or YDB MUPIP calls, we alter the REST response, so that we can analyze it and determine the validity of the test.

- Browser Caching has been enforced to prevent page caching
- Minor code clean-up in wwwroot/html/scrDashboard.html for stuff that is not used that was missed in previous commits
- Fixed dashboard storage issues
- Additional changes were implemented in the dashboard to allow the test to reach Devices table elements
- Region view: fixed bug when file was missing AND autodb = true

| File | Description |
| ---      | ---      |
|Dockerfile_test | This file is used to create a Docker image with all the packages required by the testing |
|package.json | This package is a NPM package file and describes the libraries used to perform the test and contains an entry point to execute the test |
|||
|docker-configuration/docker-test.sh | File to run testing using npm test |
|||
|test/mockData.js | This file generates mock data based on the test number |
|||
|test/node_files/bootstrap.js | This file initializes the test environment |
|test/node_files/dashboard.js | This file contains all the tests related to the `dashboard` window|
|test/node_files/devices.js | This file contains all the tests related to the `devices` dialog |
|test/node_files/libs.js | This file exports commonly used code to all and make it available to all the test files |
|test/node_files/regionView.js | This file contains all the tests related to the `region view` dialog|
|test/node_files/server.js | This file contains all the tests related to the server |
|test/node_files/systemInfo.js | This file contains all the tests related to the `server info` dialog|