[#65] M-Web-Server didn't return errors in some circumstances
For example, placing an `set X=1/0` in
getRegionStruct+157^%ydbguiRegions would cause us to have an empty
response, instead of the proper response of an error. It turned out that
the problem was that in `%ydbguiGDE` we set `$etrap` without newing it,
overriding the M Web Server's error trap. This was inadvertent and
incorrect, and is now fixed.