[#183] Remove M-Web-Server compile, install logic, delegate to M-Web-Server repo
Commit
https://github.com/shabiel/M-Web-Server/commit/45f50be210d82f7516c501a69067f046d5750eab
refactored the CMake files so that all installation for the M-Web-Server is handled by the
top level CMakeLists.txt in that project, and we only reference the
project as an external project using
FetchContent_Declare/FetchContent_MakeAvailable.

There will be future work on this, but this change is sound by itself,
as it lays down the infrastructure of how we will manage plug-in
dependencies using CMake.

How you compile/install the YottaDB GUI does not change with this
commit.

This commit also refactors the start/stop logic to be more robust,
including port open checking, and stoping an existing process when
starting a server at the same port.